+++ fili-core/src/test/groovy/com/yahoo/bard/webservice/logging/TestLogFormatter.groovy
@@ -1,0 +1,11 @@
+package com.yahoo.bard.webservice.logging
+
+/**
+ * A Test implementation of the `LogFormatter` class for the purpose of testing the LogFormatterProvider.
+ */
+class TestLogFormatter implements LogFormatter {
+    @Override
+    String format(LogBlock logBlock) {
+        return "Hello world!"
+    }
+}